+++ modules/classes.js
@@ -1,0 +1,8 @@
+export default class Book {
+    constructor(title, author, id) {
+      this.title = title;
+      this.author = author;
+      this.id = id;
+    }
+  }
+export {Book};
